requirements.txt
Pillow>=10.0.0
PyYAML>=6.0
python-dateutil>=2.8.0
pytest>=7.0.0
pytest-tmp-files>=0.0.2
